Simplify Spotlight
