+++ sentry-ruby/spec/sentry/configuration_spec.rb
@@ -258,9 +258,7 @@ RSpec.describe Sentry::Configuration do
 
   describe "#spotlight" do
     it "returns initialized Spotlight config by default" do
-      spotlight_config = subject.spotlight
-      expect(spotlight_config.enabled).to eq(false)
-      expect(spotlight_config.sidecar_url).to eq("http://localhost:8969/stream")
+      expect(subject.spotlight).to eq(false)
     end
   end
